// src/app/models/unlock-time.ts
export class UnlockTime {
     unlockTimeId: number;
     unlockTime: number;
     materialTypeId: number;
     materialTypeCode: string;
     materialTypeDesc: string;
     countryCodeId: number;
     countryCode: string;
     countryDesc: string;
     changedDateTime: string;
     userAccountChange: string;
     creationDateTime: string;
     userAccountCreation: string;
}